Add support for setting credentials via environment
This allows the user to run the build commands without neededing to interact with the cli. This is helpful for automated builds where a human is not around to enter credentials.

- I wanted a way to build my pull-requests in CI and push the artifact to a testing service such as HockeyApp.
- I wanted a way to specify different credentials or certificates based on whether or not I was building from a tagged release or a pull-request branch.

I am not super stoked by this implementation. I would rather extend `inquiry` and add support for an option such as `override` or `forceValue` where I could just pass the `process.env.VARNAME` as an option and it would essentially behave just as if a user entered the value and not show the question if the value passed the provided validator.